+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Change Snapper to only keep 10 hourly snapshots.
 - Change the NetworkManager backend to use IWD.
 - Add the ability to change the password for the "root" user during the first-time setup.
+- Add the ability to change the password for the LUKS storage encryption during the first-time setup on the secure image.
 
 ## 3.2.1 - The SteamOS 3.4 Update